split NumericBuilder into proper subclasses
