+++ std-bits/table/src/main/java/org/enso/table/data/column/operation/cast/ToFloatStorageConverter.java
@@ -1,6 +1,7 @@
 package org.enso.table.data.column.operation.cast;
 
 import org.enso.base.polyglot.NumericConverter;
+import org.enso.table.data.column.builder.object.DoubleBuilder;
 import org.enso.table.data.column.builder.object.NumericBuilder;
 import org.enso.table.data.column.storage.BoolStorage;
 import org.enso.table.data.column.storage.DoubleStorage;
@@ -32,13 +33,13 @@ public class ToFloatStorageConverter implements StorageConverter<Double> {
   }
 
   public Storage<Double> castFromMixed(Storage<?> mixedStorage, CastProblemBuilder problemBuilder) {
-    NumericBuilder builder = NumericBuilder.createDoubleBuilder(mixedStorage.size());
+    DoubleBuilder builder = NumericBuilder.createDoubleBuilder(mixedStorage.size());
     for (int i = 0; i < mixedStorage.size(); i++) {
       Object o = mixedStorage.getItemBoxed(i);
       if (o == null) {
         builder.appendNulls(1);
       } else if (o instanceof Boolean b) {
-        builder.appendDouble(b ? 1 : 0);
+        builder.appendDouble(booleanAsDouble(b));
       } else if (NumericConverter.isCoercibleToDouble(o)) {
         double x = NumericConverter.coerceToDouble(o);
         builder.appendDouble(x);
@@ -48,12 +49,12 @@ public class ToFloatStorageConverter implements StorageConverter<Double> {
       }
     }
 
-    return builder.sealDouble();
+    return builder.seal();
   }
 
   private Storage<Double> convertDoubleStorage(LongStorage longStorage) {
     int n = longStorage.size();
-    NumericBuilder builder = NumericBuilder.createDoubleBuilder(n);
+    DoubleBuilder builder = NumericBuilder.createDoubleBuilder(n);
     for (int i = 0; i < n; i++) {
       if (longStorage.isNa(i)) {
         builder.appendNulls(1);
@@ -62,21 +63,24 @@ public class ToFloatStorageConverter implements StorageConverter<Double> {
         builder.appendDouble(value);
       }
     }
-    return builder.sealDouble();
+    return builder.seal();
   }
 
   private Storage<Double> convertBoolStorage(BoolStorage boolStorage) {
     int n = boolStorage.size();
-    NumericBuilder builder = NumericBuilder.createDoubleBuilder(n);
+    DoubleBuilder builder = NumericBuilder.createDoubleBuilder(n);
     for (int i = 0; i < n; i++) {
       if (boolStorage.isNa(i)) {
         builder.appendNulls(1);
       } else {
         boolean value = boolStorage.getItem(i);
-        builder.appendDouble(value ? 1.0 : 0.0);
+        builder.appendDouble(booleanAsDouble(value));
       }
     }
-    return builder.sealDouble();
+    return builder.seal();
   }
 
+  public static double booleanAsDouble(boolean value) {
+    return value ? 1.0 : 0.0;
+  }
 }
